\name{auxiliary classes}
\docType{class}

\alias{DNAbin}
\alias{DNAbinOrNULL}
\alias{phylo}
\alias{multiPhylo}
\alias{multiPhyloOrNULL}

\alias{characterOrNULL}
\alias{numericOrNULL}
\alias{factorNULL}
\alias{integerOrNULL}
\alias{matrixOrNULL}
\alias{dataframeOrNULL}

\alias{DateOrNULL}
\alias{POSIXctOrNULL}

\title{Auxiliary classes for epibase}
\description{
  These formal (S4) classes are defined for internal use in
  \code{epibase}. They include unions of existing classes with the type
  \code{NULL}, and conversion of existing S3 classes into S4.
}
\author{Thibaut Jombart}
\keyword{classes}
