% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19_serial_interval}
\alias{covid19_serial_interval}
\title{Covid-19 Serial Intervals}
\format{
data.frame
}
\usage{
covid19_serial_interval
}
\description{
Covid-19 Serial Intervals
}
\keyword{datasets}
