\name{infectorTableToNetwork}
\alias{infectorTableToNetwork}
\title{Convert transmission tree to a network}
\usage{
  infectorTableToNetwork(transmissiontreeData)
}
\arguments{
  \item{transmissiontreeData}{Matrix of who infected whom}
}
\value{
  Network of who infected whom
}
\description{
  Convert transmission tree to a network
}
\author{
  Caroline Colijn
}

