% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_covariate_plot.R
\name{make_covariate_plot}
\alias{make_covariate_plot}
\title{Make the Covariate Plot}
\usage{
make_covariate_plot(out)
}
\arguments{
\item{out}{the output of the model}
}
\description{
Make the Covariate Plot
}
