% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_icl_plots.R
\name{make_icl_plots}
\alias{make_icl_plots}
\title{Make Imperial College London Plots}
\usage{
make_icl_plots(
  data_country,
  covariates_country_long,
  filename2,
  country,
  percent_pop,
  file_ext = "svg"
)
}
\arguments{
\item{data_country}{the country data}

\item{covariates_country_long}{the long form of covarites}

\item{filename2}{the filename}

\item{country}{the country or unit of interest}

\item{percent_pop}{the population percent}

\item{file_ext}{the file extension to write out}
}
\description{
Used internally within a function to make these plots
}
