% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_three_panel.R
\name{make_three_pannel_plot}
\alias{make_three_pannel_plot}
\title{Make Three Panel}
\usage{
make_three_pannel_plot(
  filename2,
  percent_pop = FALSE,
  covariates,
  covariates_long = NULL,
  countries,
  data_country
)
}
\arguments{
\item{filename2}{the name}

\item{percent_pop}{the population percentage}

\item{covariates}{the covariates to use}

\item{covariates_long}{if using the long covariates}

\item{countries}{the list of countries}

\item{data_country}{the data for the country}
}
\description{
Make Three Panel
}
