% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaktools-package.R
\docType{package}
\name{outbreaktools-package}
\alias{outbreaktools-package}
\alias{outbreaktools}
\title{The 'outbreaktools' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
