\name{phylofromtranstree}
\alias{phylofromtranstree}
\title{Create phylogenetic tree from transmission tree}
\usage{
  phylofromtranstree(transmissiontreeData)
}
\arguments{
  \item{transmissiontreeData}{Matrix of who infected whom}
}
\value{
  phylogenetic tree representing how samples of the
  infectious agents may be related
}
\description{
  Create phylogenetic tree from transmission tree
}
\author{
  Caroline Colijn
}

