\name{plotEpi}
\alias{plotEpi}
\title{Plot the number of susceptible, infected and recovered as a function of time}
\usage{
  plotEpi(S)
}
\arguments{
  \item{S}{Matrix containing the numbers to be plotted}
}
\description{
  Plot the number of susceptible, infected and recovered as
  a function of time
}
\author{
  Xavier Didelot
}

