% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_covariates.R
\name{process_covariates}
\alias{process_covariates}
\title{Process Covariates for Stan Models}
\usage{
process_covariates(
  countries,
  interventions,
  d,
  ifr.by.country,
  N2,
  serial_interval = NULL,
  min_date = "31/12/2019"
)
}
\arguments{
\item{countries}{the countries or units to consider}

\item{interventions}{the data.frame of interventions}

\item{d}{the data.frame of interventions}

\item{ifr.by.country}{IFR by country}

\item{N2}{number of simulations}

\item{serial_interval}{the path to serial intervals, use default}

\item{min_date}{the minimum date to use in the padding}
}
\description{
Process Covariates for Stan Models
}
