\name{sciencegallery}
\docType{data}
\alias{sciencegallery}

\title{Infectious SocioPatterns dynamic contact networks}
\description{
This dataset contains the dynamic contact network collected during the Infectious SocioPatterns event that took place at the Science Gallery in Dublin, Ireland, during the artscience exhibition INFECTIOUS: STAY AWAY. The data represent the active contacts during 20-second intervals of one day of data collection; this is a subset of the full data, from 2009-04-28. The data comprise of an anonymous IDs for pairs of individuals, and the time stamp (in POSIXct) during which the contact was active. The data were collected by the SocioPatterns collaboration (www.sociopatterns.org).
}

\references{Isella et al. (2011). "What’s in a crowd? Analysis of face-to-face behavioral networks." Journal of Theoretical Biology 271:166.}
\keyword{dataset}
