\name{singapore}
\docType{data}
\alias{singapore}

\title{Dataset from a small SARS outbreak in Singapore in 2003}
\description{
These are the different datasets for the outbreak:
singapore: a file with 14 sequences with dates
phylogeny: the phylogeny as recontructed in Liu et al. PLoS Med 2005
transmissionNetwork: a possible transmission tree inferred from the phylogeny reconstructed
}
\references{Liu, J., et al. (2005). "SARS transmission pattern in Singapore reassessed by viral sequence variation analysis." PLoS Med 2(2): e43.}
\keyword{dataset}
