% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icl_stan.R
\name{stan_icl_general}
\alias{stan_icl_general}
\title{Stan Model Base General}
\usage{
stan_icl_general(
  dat,
  iter = 2400,
  warmup = 1000,
  chains = 4,
  control = list(adapt_delta = 0.95, max_treedepth = 13),
  ...
)
}
\arguments{
\item{dat}{the dat sent to the model}

\item{iter}{2400}

\item{warmup}{1000}

\item{chains}{4}

\item{control}{default controls}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Stan Model Base General
}
