\name{auxiliary classes}
\docType{class}

\alias{DNAbin-class}
\alias{DNAbinOrNULL}
\alias{DNAbinOrNULL-class}
\alias{phylo-class}
\alias{multiPhylo-class}
\alias{multiPhyloOrNULL}
\alias{multiPhyloOrNULL-class}

\alias{characterOrNULL}
\alias{characterOrNULL-class}
\alias{numericOrNULL}
\alias{numericOrNULL-class}
\alias{factorNULL}
\alias{factorNULL-class}
\alias{integerOrNULL}
\alias{integerOrNULL-class}
\alias{matrixOrNULL}
\alias{matrixOrNULL-class}
\alias{data.frameOrNULL}
\alias{data.frameOrNULL-class}
\alias{listOrNULL}
\alias{listOrNULL-class}

\alias{DateOrNULL}
\alias{DateOrNULL-class}
\alias{POSIXctOrNULL}
\alias{POSIXctOrNULL-class}

\title{Auxiliary classes for OutbreakTools}
\description{
  These formal (S4) classes are defined for internal use in
  \code{OutbreakTools}. They include unions of existing classes with the type
  \code{NULL}, and conversion of existing S3 classes into S4.
}
\author{Thibaut Jombart}
\keyword{classes}
