\name{fake_SARS_HK}
\docType{data}
\alias{fake_SARS_HK}

\title{fake dataset similar to the SARS outbreak in Hong Kong in 2003}
\description{
  This data set is a fake data set that was designed to reproduce interesting features of the SARS (severe acute respiratory syndrom) outbreak in Hong Kong in 2003. It has a similar size and dynamics, as well as realistic breakdown of cases between cases infected in Prince of Wales hospital (where a large super spreading event (SSE) occurred, exposure code 1), Amoy Gardens residence (where another SSE occurred, exposure code 2), cases infected in other hospitals (exposure code 3), and cases with unknown source of infection (exposure code 4). The exposure start and exposure end fields are present for cases who are known to have been in contact with known cases during a certain period of time defined by those two dates.   
}
\keyword{dataset}
