\name{HorseFluOutbreak}
\docType{data}
\alias{HorseFluOutbreak}

\title{Dataset from the Newmarket 2003 equine Influenza outbreak}
\description{
These are the different datasets for the outbreak:
HorseFluSeq: a FASTA sequence file with approx 2000 sequences obtained by cloning and Sanger sequencing (accession number, nucleotide sequence)
HorseFluYardInfo: a file with the geo-location of each yard (yard identifier, decimal longitude, decimal latitude)
HorseFluSeqID: Information about the sample that each sequence was taken from, multiple samples were taken from a few horses. (Accession, tube identidier, clone identifier)
HorseFluShedding: Data on the viral shedding load for each sample (Tube identifier , viral shedding load, yard identifer, horse identifier, swad date)
HorseFluHorseInfo: Limited information relating to the host (age, sex, date of first vaccinaton)

}
\references{
Hughes J, Allen RC, Baguelin M, Hampson K, Baillie GJ, et al. (2012) Transmission of Equine Influenza Virus during an Outbreak Is Characterized by Frequent Mixed Infections and Loose Transmission Bottlenecks. PLoS Pathog 8(12): e1003081. doi:10.1371/journal.ppat.1003081
}
\keyword{dataset}
