\name{singapore}
\docType{data}
\alias{singapore}

\title{Dataset from a small SARS outbreak in Singapore in 2003}
\description{
This dataset contains four objects:

\code{individuals} is a dataframe containing individual identifiers and the ID of their potential infector, according to the phylogeny inferred by Liu et al. PLoS Med 2005. 
	
\code{samples} is a dataframe containing individual IDs, sample IDs, date the sample was collected and ID of a sequence taken from that sample. 
	
\code{dna} contains 13 alined whole genome sequences refferred to in \code{samples}. 

\code{trees} is a list containing the phylogeny reconstructed by Liu et al. PLoS Med 2005.
}
\references{Liu, J., et al. (2005). "SARS transmission pattern in Singapore reassessed by viral sequence variation analysis." PLoS Med 2(2): e43.}
\keyword{dataset}
\author{
  Anne Cori, Xavier Didelot, Joseph Hughes
}